refactor(native): extract StellarTokenActions component for better readability
